apps/api/pyproject.toml
[project]
name = "videogen-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.0",
  "SQLAlchemy>=2.0.25",
  "alembic>=1.13.0",
  "asyncpg>=0.29.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "httpx>=0.27.0",
  "redis>=5.0.0",
  "boto3>=1.34.0",
  "botocore>=1.34.0",
  "tenacity>=8.2.3",
  "stripe>=8.0.0"
]

[tool.alembic]
script_location = "apps/api/app/db/migrations"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

